send_sms dies with error message if unable to send
If send_sms is not able to deliver the message for whatever reason, the
call now dies with a helpful error message.

This is as specified in SMS::Send documentation.
Previously, the script did NOT die and no value was returned.
Also, there was no indication of what caused the failure.