Cannot escape utf8 characters #7
We are getting errors like this:

Can't escape \x{014D}, try uri_escape_utf8() instead at /usr/local/share/perl/5.28.1/WWW/Twilio/API.pm line 116.

It appears that WWW::Twilio::API has a param "utf8" for new() that would cause the module to use uri_escape_utf8, but there doesn't appear to be a way to enable it in SMS::Send::Twilio.
If it could be added as a parameter to new() that would be great!